MappedFile问价获取最大读指针/销毁、TransientStorePool属性、CommitLog文件、ConsumeQueue文件

+++ store/src/main/java/org/apache/rocketmq/store/config/MessageStoreConfig.java
@@ -21,11 +21,11 @@ import org.apache.rocketmq.common.annotation.ImportantField;
 import org.apache.rocketmq.store.ConsumeQueue;
 
 public class MessageStoreConfig {
-    //The root directory in which the log data is kept
+    // The root directory in which the log data is kept
     @ImportantField
     private String storePathRootDir = System.getProperty("user.home") + File.separator + "store";
 
-    //The directory in which the commitlog is kept
+    // The directory in which the commitlog is kept
     @ImportantField
     private String storePathCommitLog = System.getProperty("user.home") + File.separator + "store"
         + File.separator + "commitlog";
@@ -199,7 +199,7 @@ public class MessageStoreConfig {
     public int getMappedFileSizeConsumeQueue() {
 
         int factor = (int) Math.ceil(this.mappedFileSizeConsumeQueue / (ConsumeQueue.CQ_STORE_UNIT_SIZE * 1.0));
-        return (int) (factor * ConsumeQueue.CQ_STORE_UNIT_SIZE);
+        return factor * ConsumeQueue.CQ_STORE_UNIT_SIZE;
     }
 
     public void setMappedFileSizeConsumeQueue(int mappedFileSizeConsumeQueue) {
